requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.1.0
discord==2.2.3
discord.py==2.2.3
distlib==0.3.6
filelock==3.12.0
frozenlist==1.3.3
hydra-core==1.3.2
identify==2.5.24
idna==3.4
multidict==6.0.4
nodeenv==1.8.0
omegaconf==2.3.0
openai==0.27.6
packaging==23.1
platformdirs==3.5.1
pre-commit==3.3.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.30.0
tqdm==4.65.0
urllib3==2.0.2
virtualenv==20.23.0
yarl==1.9.2
